+++ sdk/compute/arm-compute/src/models/parameters.ts
@@ -25,6 +25,7 @@ import {
   VirtualMachineScaleSetVMExtensionUpdate as VirtualMachineScaleSetVMExtensionUpdateMapper,
   VirtualMachineScaleSetVMReimageParameters as VirtualMachineScaleSetVMReimageParametersMapper,
   VirtualMachineScaleSetVM as VirtualMachineScaleSetVMMapper,
+  AttachDetachDataDisksRequest as AttachDetachDataDisksRequestMapper,
   RunCommandInput as RunCommandInputMapper,
   VirtualMachineExtension as VirtualMachineExtensionMapper,
   VirtualMachineExtensionUpdate as VirtualMachineExtensionUpdateMapper,
@@ -110,7 +111,7 @@ export const $host: OperationURLParameter = {
 export const apiVersion: OperationQueryParameter = {
   parameterPath: "apiVersion",
   mapper: {
-    defaultValue: "2023-03-01",
+    defaultValue: "2023-07-01",
     isConstant: true,
     serializedName: "api-version",
     type: {
@@ -424,6 +425,11 @@ export const sasUriExpirationTimeInMinutes: OperationQueryParameter = {
 
 export const parameters5: OperationParameter = {
   parameterPath: "parameters",
+  mapper: AttachDetachDataDisksRequestMapper
+};
+
+export const parameters6: OperationParameter = {
+  parameterPath: "parameters",
   mapper: RunCommandInputMapper
 };
 
@@ -460,17 +466,17 @@ export const extensionParameters5: OperationParameter = {
   mapper: VirtualMachineExtensionUpdateMapper
 };
 
-export const parameters6: OperationParameter = {
+export const parameters7: OperationParameter = {
   parameterPath: "parameters",
   mapper: VirtualMachineCaptureParametersMapper
 };
 
-export const parameters7: OperationParameter = {
+export const parameters8: OperationParameter = {
   parameterPath: "parameters",
   mapper: VirtualMachineMapper
 };
 
-export const parameters8: OperationParameter = {
+export const parameters9: OperationParameter = {
   parameterPath: "parameters",
   mapper: VirtualMachineUpdateMapper
 };
@@ -505,7 +511,7 @@ export const expand4: OperationQueryParameter = {
   }
 };
 
-export const parameters9: OperationParameter = {
+export const parameters10: OperationParameter = {
   parameterPath: ["options", "parameters"],
   mapper: VirtualMachineReimageParametersMapper
 };
@@ -612,7 +618,7 @@ export const typeParam: OperationURLParameter = {
   }
 };
 
-export const parameters10: OperationParameter = {
+export const parameters11: OperationParameter = {
   parameterPath: "parameters",
   mapper: AvailabilitySetMapper
 };
@@ -628,12 +634,12 @@ export const availabilitySetName: OperationURLParameter = {
   }
 };
 
-export const parameters11: OperationParameter = {
+export const parameters12: OperationParameter = {
   parameterPath: "parameters",
   mapper: AvailabilitySetUpdateMapper
 };
 
-export const parameters12: OperationParameter = {
+export const parameters13: OperationParameter = {
   parameterPath: "parameters",
   mapper: ProximityPlacementGroupMapper
 };
@@ -649,7 +655,7 @@ export const proximityPlacementGroupName: OperationURLParameter = {
   }
 };
 
-export const parameters13: OperationParameter = {
+export const parameters14: OperationParameter = {
   parameterPath: "parameters",
   mapper: ProximityPlacementGroupUpdateMapper
 };
@@ -664,7 +670,7 @@ export const includeColocationStatus: OperationQueryParameter = {
   }
 };
 
-export const parameters14: OperationParameter = {
+export const parameters15: OperationParameter = {
   parameterPath: "parameters",
   mapper: DedicatedHostGroupMapper
 };
@@ -680,12 +686,12 @@ export const hostGroupName: OperationURLParameter = {
   }
 };
 
-export const parameters15: OperationParameter = {
+export const parameters16: OperationParameter = {
   parameterPath: "parameters",
   mapper: DedicatedHostGroupUpdateMapper
 };
 
-export const parameters16: OperationParameter = {
+export const parameters17: OperationParameter = {
   parameterPath: "parameters",
   mapper: DedicatedHostMapper
 };
@@ -701,7 +707,7 @@ export const hostName: OperationURLParameter = {
   }
 };
 
-export const parameters17: OperationParameter = {
+export const parameters18: OperationParameter = {
   parameterPath: "parameters",
   mapper: DedicatedHostUpdateMapper
 };
@@ -734,7 +740,7 @@ export const hostName1: OperationURLParameter = {
   }
 };
 
-export const parameters18: OperationParameter = {
+export const parameters19: OperationParameter = {
   parameterPath: "parameters",
   mapper: SshPublicKeyResourceMapper
 };
@@ -750,12 +756,12 @@ export const sshPublicKeyName: OperationURLParameter = {
   }
 };
 
-export const parameters19: OperationParameter = {
+export const parameters20: OperationParameter = {
   parameterPath: "parameters",
   mapper: SshPublicKeyUpdateResourceMapper
 };
 
-export const parameters20: OperationParameter = {
+export const parameters21: OperationParameter = {
   parameterPath: "parameters",
   mapper: ImageMapper
 };
@@ -771,12 +777,12 @@ export const imageName: OperationURLParameter = {
   }
 };
 
-export const parameters21: OperationParameter = {
+export const parameters22: OperationParameter = {
   parameterPath: "parameters",
   mapper: ImageUpdateMapper
 };
 
-export const parameters22: OperationParameter = {
+export const parameters23: OperationParameter = {
   parameterPath: "parameters",
   mapper: RestorePointCollectionMapper
 };
@@ -792,7 +798,7 @@ export const restorePointCollectionName: OperationURLParameter = {
   }
 };
 
-export const parameters23: OperationParameter = {
+export const parameters24: OperationParameter = {
   parameterPath: "parameters",
   mapper: RestorePointCollectionUpdateMapper
 };
@@ -807,7 +813,7 @@ export const expand5: OperationQueryParameter = {
   }
 };
 
-export const parameters24: OperationParameter = {
+export const parameters25: OperationParameter = {
   parameterPath: "parameters",
   mapper: RestorePointMapper
 };
@@ -833,7 +839,7 @@ export const expand6: OperationQueryParameter = {
   }
 };
 
-export const parameters25: OperationParameter = {
+export const parameters26: OperationParameter = {
   parameterPath: "parameters",
   mapper: CapacityReservationGroupMapper
 };
@@ -849,7 +855,7 @@ export const capacityReservationGroupName: OperationURLParameter = {
   }
 };
 
-export const parameters26: OperationParameter = {
+export const parameters27: OperationParameter = {
   parameterPath: "parameters",
   mapper: CapacityReservationGroupUpdateMapper
 };
@@ -874,7 +880,7 @@ export const expand8: OperationQueryParameter = {
   }
 };
 
-export const parameters27: OperationParameter = {
+export const parameters28: OperationParameter = {
   parameterPath: "parameters",
   mapper: CapacityReservationMapper
 };
@@ -890,7 +896,7 @@ export const capacityReservationName: OperationURLParameter = {
   }
 };
 
-export const parameters28: OperationParameter = {
+export const parameters29: OperationParameter = {
   parameterPath: "parameters",
   mapper: CapacityReservationUpdateMapper
 };
@@ -905,12 +911,12 @@ export const expand9: OperationQueryParameter = {
   }
 };
 
-export const parameters29: OperationParameter = {
+export const parameters30: OperationParameter = {
   parameterPath: "parameters",
   mapper: RequestRateByIntervalInputMapper
 };
 
-export const parameters30: OperationParameter = {
+export const parameters31: OperationParameter = {
   parameterPath: "parameters",
   mapper: ThrottledRequestsInputMapper
 };
@@ -1349,22 +1355,22 @@ export const roleName: OperationURLParameter = {
   }
 };
 
-export const parameters31: OperationParameter = {
+export const parameters32: OperationParameter = {
   parameterPath: ["options", "parameters"],
   mapper: CloudServiceMapper
 };
 
-export const parameters32: OperationParameter = {
+export const parameters33: OperationParameter = {
   parameterPath: ["options", "parameters"],
   mapper: CloudServiceUpdateMapper
 };
 
-export const parameters33: OperationParameter = {
+export const parameters34: OperationParameter = {
   parameterPath: ["options", "parameters"],
   mapper: RoleInstancesMapper
 };
 
-export const parameters34: OperationParameter = {
+export const parameters35: OperationParameter = {
   parameterPath: ["options", "parameters"],
   mapper: UpdateDomainMapper
 };